plat: Tweak Cypress code for Arm Compiler version 6
Arm compiler version 6 doesn't support using __asm to
declare a function in C code. For now, check the compiler
version and just don't define Cy_MemorySymbols() or
Cy_SysLib_AsmInfiniteLoop() if we're building with that
compiler. If we find that we need these functions,
we can add them back in a suitable form at a later date.

zero_init is no longer supported in a section attribute, so
migrate as described in
http://www.keil.com/support/man/docs/armclang_mig/armclang_mig_chr1398848377314.htm

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: I0f33af72c551eb7c637ee5df220b587546d94b94