platform: Add static protection for PSoC6
Configure SMPUs and PPUs and set Protection Contexts for
all bus masters.
The SMPU protection does leave some gaps - the current memory
map is awkward to protect and is likely to change going forward
(in particular due to the change to use overwrite-only mode
for image upgrade).
A lot of driver functions are not yet implemented.

Change-Id: I6ecebf3c82ca559a5f966fbd1e5f7f12d29834df