platform: Add PSoC hard fault handler on Cortex-M4
Similar to Cypress PDL 3.0.1.

This adds two weak symbols:
Cy_SysLib_FaultHandler is an infinite loop.
HardFault_Handler is the Cypress PDL handler for a hard fault,
that stores LR if needed and jumps to Cy_SysLib_FaultHandler.

The weak infinite loop HardFault_Handler that was previously
present is removed.

This change by itself has little effect, but it does mean that
a (non-weak) Cy_SysLib_FaultHandler, if defined, will be called
when a hard fault occurs.

Note that while there is a HardFault_Handler defined by TF-M, it
is only built as part of the secure SPE build, not as part of the
non-secure host build so it has no effect here.

Change-Id: I5b1a8c6cc8f80e0fb86e2a57c274ce5f1f32dc20
Signed-off-by: Chris Brand <chris.brand@cypress.com>