Twincpu: Add updated common linker scatter file
Commit 4ca5c66e26aa86486fb514db00858766cce24996
("Twincpu: Add preprocessor option and configuration
TFM_MULTI_CORE_TOPOLOGY") updated the file
platform/ext/common/armclang/tfm_common_s.sct.template
but did not include the updated generated file, which is also
under source control for some reason. As the actual scatter file
is not generated as part of the build, this forces the user to
manually re-generate it before building.

Rectify that omission.

Change-Id: Ia06c9926ea8a54a8693deaab63c5ac73ea33f3a0
Signed-off-by: Chris Brand <chris.brand@cypress.com>