RSE: add rse persistent data
Add persistent region to RSE for storing data expected to survive a cold
reset.

The persistent data region is for information that is expected to
persist resets. This is not expected to have a value on a cold boot.
This can be used to communicate between resets such as when debug is
enabled or whether the last boot was successful.

Change-Id: Ifc0727273ebcd2e953aa92ade0fb5f16ef2fe75e
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>