RSE: Add provisioning status reports
Update provisioning to use the DCSU REPORT_STATUS command to report
provisioning errors and milestones.

The message is split into two parts a type and a value.

Type indicates how the message should be processed by tooling.
* SUCCESS CONTINUE indicates an operation was successful and more
  messages will follow. The value indicates what the operation was.
* SUCCESS COMPLETE indicates an operation was successful and tooling
  can stop waiting.
* ERROR indicates and operation failed and tooling can stop waiting. The
  value indicates the error code.

Blob validation and several provisioning blob milestones such as TP Mode
set and BL1-2 fuses blown are reported along with any errors that occur
in provisioning handling.

If the provisioning was not started by DCSU then the messages will not
be sent.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I06130374f093645c39bc218282f38d98776f3726
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>