LIB: Move partitions to use new logging API
Make use of the new unprivileged logging API throughout the secure
partitions. This mainly does drop in replacements for the existing
macros to the new LOG_UNPRIV macros.

Note that as part of this change, we replace the existing remove
tfm_sp_log_raw.c which implemented the vprintf standard library
function. This function is still by some platforms and external
libraries in TF-M and therefore we add a new implementation which makes
use of the implementation in the new unpriv logging library.

Change-Id: I58c7f71fabab0984fe42560c6c20ca313bd5aa60
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>