RSE: DCSU separate RX status from error
DCSU was using the error return to set the RX status return. This
complicated the response system and error system in rse_dcsu_hal due to
the number of conversions that error and status messages went through.

To resolve this separate the status reports from error reports. This
will allow commands to return a status report without returning it as
an error. This will reserve the error return for errors that would
prevent receiving further commands.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I59745583409eb8db756c0da1293589b9e847e264
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>