RSE: Support for passing HUK buffer for IAK seed generation
The IAK seed needs to be generated during CM provisioning. It
derives from the HUK but during CM provisioning the KMU slot
containing the HUK is not populated until the next reset. For
this reason we extend the function that derives the IAK seed
to receive the HUK as raw input bytes which are available
during CM provisioning.

Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Co-authored-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I88784357061a8cadf9ffc864f5c011d574946fdd