LIB: Use tfm_log for BL2 logging
Make use of the new tfm_log library for logging from BL2. We achieve
this by redefining the MCUBOOT_LOG_ macros in TF-M to use the tfm_log
macros instead of printf. The log level is mapped from the configured
MCUBoot value to the value used by tfm_log and we add a CPP check to
ensure that the two match correctly.

Change-Id: I6dd24bd53008a779ab5fa282153a8c8d5c4e5741
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>