Platform: Add custom __assert_func for BL1/2
This patch adds a definition for __assert_func, which is the function
called if an assertion fails in C. We redefine the default function to
instead use the tfm_log library, therefore removing the dependency on
libc. After logging the issue, the function just spins in a while (1).

Note that because of the removed dependency, this results in a
significant size reduction for BL1_1 and BL2.

Change-Id: I79f514d92fcd1766fea9c464aff09b1f5832984a
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>