SPRT: Add custom __assert_func
Add a custom implementation of the __assert_func for the secure runtime
code. This is the function called when an assertion fails (when using
assert from the standard header).

This implementation is placed in the shared runtime code so that it can
be used by both privileged and unprivileged code. It uses a check to
differentiate between the two and call the appropriate logging function
before entering a while (1).

Change-Id: If4c26e944ae40533b2f8ab6d63cda8fda088563d
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>