SPM: Remove custom assert.h
Remove the custom implementation of assert.h for the SPM and move any
files using SPM_ASSERT to use plain old assert. We now have custom
implementations of __assert_func (the assert implementation) for each of
the BL components and the secure firmware which follows the existing
implementation of SPM_ASSERT.

Note that as part of this change, we can remove the workaround added to
tfm_log.h to #ifdef the VERBOSE macro. This was required as the
psa_arch_tests were including the custom assert.h and pulling in
tfm_log.h. Now that the custom assert.h is removed, this is no longer an
issue.

Change-Id: I69fd3093633c5b5f502897540f477387392b12eb
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>