RSE: DCSU Remove clear synchronization
This patch removes DCSU clear synchronization that was being used to
ensure both COMMAND and STATUS were clear before the next command was
received.

DCSU should not require additional synchronization as setting the STATUS
to a non zero value should clear the COMMAND and setting the COMMAND to
a non zero value should clear the STATUS. This means any additional
synchronization around clearing these values is unnecessary.

The only reason this was required before is a bug in FVP.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: Ifa2b954411e81a0e05ae929281f2c3ac63649ba1
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>