RSE: Add DCSU driver tests
Add a test suite for the DCSU driver. This suite consists of two parts,
the RSE tests which are registered with the extra BL1_1 tests in a
similar way to the CC3XX tests, and a python script which exercises the
DCSU.

This python script imports from the main DCSU script and uses it to
send and receive commands from the DCSU test suites. It should be run at
the same time as the tests are being run. Each of the tests it provides
can pass or fail, but it is worth noting that both the RSE side test and
host side test must pass in order for there to be an actual test pass.

The addition of new RSE tests also increases the size of the RSE test
binary and therefore we need to increase its code size accordingly.

Change-Id: Ied71d6a3a1cafe67cc87f80831429f06576cb223
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>