RSE: Specify explicit BL1_1_DATA_SIZE
Previously half of the remaining DTCM size after removing the
BOOT_TFM_SHARED_DATA was allocated to BL1_1 and BL1_2 data. Change this
instead to give the BL1_1 an explicit size for its data and give BL1_2
whatever is left over.

Note that we also add the BL1_1 and BL1_2 to the size preprocessor
checks. This ensures that the calculations we have added are valid.

Change-Id: I9de7150893abf9711bd7e13016f5224fbc243e2e
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>