Build: Move tf-m-extras version to base config
Move the definitions of the tf-m-extras version to checkout to the
config_base.cmake file. This ensures that all platforms share this CMake
variable instead of it only being set for RSE as was the case
previously.

Note that platforms can continue to override these variables with a
specific version if they need to, or can set the variables as empty if
they do not want the tf-m-extras repo to be checked out at all.

Change-Id: I71eaf75044ba58d132af70f0556e62fb86fef774
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>