LIB: Partially rework BL2 logging
Instead of verifying that the LOG_LEVEL matches the MCUBOOT_LOG_LEVEL,
completely ignore the definition of LOG_LEVEL for MCUBoot logging and
ensure that instead we only use the MCUBOOT_LOG_LEVEL macro to determine
the level to log at.

We achieve this by setting the LOG_LEVEL to LOG_LEVEL_VERBOSE which
compiles in all logging macros and allow the MCUBOOT_LOG_LEVEL to set
which level to log at.

Change-Id: I343774306734499b691fde04cd87626ff6d13add
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>