Build: Also strip workspace from __FILE__
Add another -fmacro-prefix-map argument to the GCC/Clang command line
which strips the name of the workspace from the __FILE__ macro. This
applies primarily to files which live outside of TF-M and will ensure
that the following transformations now occur:

/workspace/tf-m-tests/directory/file.c -> tf-m-tests/directory/file.c
/workspace/trusted-firmware-m/directory/file.c -> directory/file.c

This ensures that we remove user specific elements of the build which
also keeping assertions traceable (as we can easily see source files
that are located outside of trusted-firmware-m).

Change-Id: I26cd1e0049120ba108524ddedaed6b6b69f60dcf
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>