Build: Add optimisation flag build option
Add a TFM wide flag which allows for adding -Og optimisation for Debug
builds. Note that this does not have an impact on assertions which will
still be enabled in Debug builds and will not effect any other command
line flags, it only passes -Og.

This optimisation flag has the following benefits:
- Speeds up code execution
- Vastly reduces code size
- Allows the optimiser to generate more advanced warnings such as
-Wmaybe-uninitialised

The drawback is that optimisations make the code more difficult to step
through although with -Og, given that this flag is designed for
debugging, these should be minimal. The user will still
be able to build without optimisations if desired by overriding this
flag on the command line.

Change-Id: I93e7603700b3111e62723c72087211bbe080a40d
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>