RSE: Add expansion interrupt macro
Add a macro which allows platforms to define handlers for expansion
interrupts and for them to be placed in the vector table.

Note that as part of this commit, we remove the definition of
S_CODE_VECTOR_TABLE_SIZE. This causes the linker script to place the
text section after the vector table regardless of its size.

We also fix the indentation of the vector table and other macros in this
file to make them consistent with the 4 space indent coding style.

Change-Id: I30be63c30be119d572ed2abcb30da9c6a5b4af2c
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>