RSE: DCSU: Fix EXPORT_DATA command
Fix the handling of this command in the RSE and dcsu.py script. We do
this by changing the diag_tx_large_param register to instead specify the
offset in the buffer to write to. The RSE will then send a
COMPLETE_EXPORT_DATA command to indicate that it is done sending data.

Change-Id: I9ed6cc0ca8c0bf730a559aa9eb5fde294121f5dd
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>