RSE: Force IRQn_Type size
Previously, the IRQn_Type enum only specified values up to 62 and
therefore, because we compile with -fshort-enum and the enum contains
negative values, the enum type will be a signed char. When we try and
use the expansion interrupts, they will be truncated to negative values.

Add padding to the enum to force its size to be int32_t.

Change-Id: I57c46b894748ffa71df2b64e93d75388c88ca0e3
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>