RSE: Allow DM provisioning in BL1_2
Add tfm_plat_provisioning_ function implementations for BL1_2 which are
then called in the boot_hal_bl1_2.c. The functions are essentially the
same as the BL1_1 functions but without the RSE persistent data
handling. The check for whether provisioning is required is also changed
so that provisioning is only performed in BL1_2 when the
RSE_BOOT_IN_DM_LCS flag is set and we are in DM LCS.

Note that we call only try and perform provisioning in BL1_2 in the case
that the flash load fails. Otherwise we continue booting up to the
runtime firmware which will handle the RSE provisioning instead.

Change-Id: Id9108c6800acd6a8985874ee94c1b7d904268048
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>