RSE: otp_lcm: Fix multi-RSE mappings
Fix the OTP mappings table entries for the RSE_ID and RSE to RSE routing
tables. In the case of the RSE_ID, the entry was missing and should be
defined in the case that it is present in the OTP. The routing table
entires were invalid and do not compile if the
RSE_OTP_HAS_ROUTING_TABLES flag is set.

Change-Id: I582509cf5fede1df55e779ac58511c40e10105b0
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>