RSE: Share stack and heap between BL1_1 and BL1_2
Export the stack and heap linker script variables from
BL1_1 to BL1_2 and remove the definitions from the BL1_2 linker script.
This solution allows us to save space in BL1_2.

Note that we have to remove the PROVIDE keyword in the __stack
definition in rse_tests.ld. This is because the PROVIDE keyword will
cause the __stack symbol not to be overridden and thus the RSE tests
will use the BL1_1 stack.

Change-Id: I83790e830cf75b64d432354f704557d9522a7bf7
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>