RSE: Rework provisioning blob unpack locations
Fix the size of the BL1_2 data section (this is now vastly reduced given
that it shares the stack and heap from BL1_1). With this fixed size, we
can now allocate all of the remaining space for the provisioning blob
data.

This patch also moves the code location to the beginning of the ITCM and
the values location to the beginning of the DTCM.

Change-Id: I459a0486c5028922d98a786023ece7d12d000f1a
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>