cc3xx: support CMAC KDF key sizes not aligned to AES block size
The CMAC KDF is currently limited to generating keys which size is a
multiple of AES block size (16 bytes).

This eliminates such limitation by allowing unaligned key sizes as long
as they are 32-bit aligned, and comply with NIST SP800-38B which in this
case means the last unaligned block should be at least 8 bytes long.

This is tested for:
- 128-bit key.
- 192-bit key.
- 224-bit key.
- 256-bit key.

Change-Id: I2d70e45208796ea0eb919d66765f7d8390c2807d
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>