Crypto: Do not error out if failed to load key
Instead of exiting the tfm_builtin_key_loader_init function if we fail
to load a single key, continue in the loop without setting the is_loaded
field. When we come to use the key, we check that this field is set and
therefore we will fail at that point.

Change-Id: Idc19b5a0474b7385ca27715571176f5ca6ddc5c1
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>