RSE: Fix ROTPK revocation and add to build
Fix the following issues with the ROTPK revocation handling functions:
- Missing void in function declarations
- Missing platform error enum definitions
- CM revocation handling did not work as expected. This is due to the
P_RSE_OTP_CM_ROTPK pointer not being initialised when the zero count is
incorrect. This will be the case once we have revoked the current ROTPKs
and therefore we cannot use it to write the new ROTPKs. We therefore
do not use the P_RSE_OTP_CM_ROTPK macro directly and instead use the
P_RSE_OTP_CM pointer

As part of this patch we also add this file to the secure runtime build.
None of the functions are currently used but this prevents the file from
becoming out of date.

Change-Id: I5cb512a43d10444f1b824a5cd3335fbe6fb5d981
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>