RSE: Add non endorsed plain data handler
Add a function which can be passed in the context structure of
handle_provisioning_message to parse and load a plain data message. This
function can initially only handle a single type of plain data - non
endorsed provisioning data, which contains a structure of the type
struct rse_non_endorsed_dm_provisioning_values_t which is also added as
part of this commit.

Change-Id: I9127d1532a7199a80b8442bb92c070e62320b5bd
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>