RSE: respond with bus error to security violation
SECRESPCFG is a signal which configures how to respond to an access when
a security violation occurs. 0x0U for RAZ / WI, 0x1U for bus error.

This is currently disabled until TF-M runtime stage, which may hide
security violations in earlier stages.

Configure SECRESPCFG to raise a bus error starting from BL1.

Change-Id: I7451a7ff0f5cf4d15175b6177605694edd07b748
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>