RSE: Add script to generate non endorsed data
Add additional python script to support non endorsed DM provisioning and
also make updates to the existing provisioning_config.py script to
support this.

The main change to provisioning_config.py is to allow it to process
provisioning value structures without multiple ROTPK areas (as is the
case with the non endorsed provisioning values).

Change-Id: I0ac19161f2ddfb670b82a717770c296b02fd2927
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>