RSE: Split message and blob handler
Split out the blob handling function into its own source file. This
allows it to be compiled separately (as it will not be required for
runtime). This patch also changes the blob status functions to make them
more generic and allow them to be used for reporting statuses for other
types of messages.

Change-Id: Ied3203686dd49f72b76c5697469f8ce8a2829a10
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>