RSE: Fix warnings on LCM driver, BL11 boot HAL and FIP parser
Fix warnings related to the usage of init_atu_regions() in the
BL11 boot HAL which is conditional on some build options, and
cover all possible cases in a LCM driver switch case, even if
the case we know is unreachable, cover it explicitly. Use proper
ifdef guards in the FIP parser with functions used when GPT
support is required.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I254691070ed06acb51e61c37aa017d2e708a43e3