Platform: Corstone1000: Remove duplicate configuration parameters for Corstone-1000
The PS_NUM_ASSET is duplicated in the cmake.config and the
config_tfm_target.h file under Corstone-1000. The commit removes
the one from the cmake.config and keeps the one in the header file.

The whole rationale behind this is for the vendor to be able
to override the configuration using the cmake file.

Change-Id: I5a36924386d831a534d5ee17e16814a69996a8a9
Signed-off-by: Yogesh Wani <yogesh.wani@arm.com>