RSE: Add plain data interpreting blob
Add source file and python tooling to generate a blob which interprets a
plain data message. Also add support for running this blob in the BL1_2
blob provisioning code; as with the BOOT_IN_DM_LCS provisioning, this
code is run in the case that the flash loading fails.

The CMake configuration required to support this blob is the same as the
ROTPK in blob flow, as in this case we also use a public key embedded in
the blob to validate the blob before running it.

Change-Id: I160af5a082a31f0c83942be22b59fb518eb2bf9a
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>