RSE: Tidy up provisioning config
Split the provisioning configuration variables between symmetric (AES)
provisioning and asymmetric (AES with ECDSA). Putting all of the
possible configuration values in one of these two categories clarifies
the configuration and makes it less prone to errors.

Note that we also update some of the doc strings for the variables to
clarify their purpose.

Change-Id: Ifd4ad20a1eb2a89b4328e903690c3e6900f0fd98
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>