RSE: Use cm_area_info from shared symbols
Previously, the otp_lcm.c file was compiled separately for BL1_2 and
therefore it was using a separate set of symbols from BL1_1. As BL1_2
did not call tfm_plat_otp_init, these symbols were never configured and
are therefore 0 on startup. This cause the P_RSE_OTP_CM pointer to be
NULL and therefore when trying to access P_RSE_OTP_CM->config_flags, we
were actually reading from address 0x8 in memory (which is the offset of
the config_flags field).

Remove the separately compiled otp_lcm.c and add cm_area_info to the RSE
shared symbols.

Change-Id: I66a50712f72dfd20f6b049cf38d5503934fbf8db
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>