RSE: Add OTP pointer check and get functions
Add a getter inline function for the area info pointer and use in each
of the area pointer macros. This allows for better type checking and
also to add an assertion to the pointer check.

Also add a function to allow for checking if the OTP area is valid
before attempting to use the pointer. This is expected to be used in
cases where the area may not be valid and we want to check without
asserting.

Change-Id: Idf10b657f996a0487c5b7a482db746481bbcf8cd
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>