platform: nordic_nrf: Add configuration for GPIOTE security
In nRF54L series the GPIOTE peripheral has user selectable
security configuration.
In addition to this the GPIOTE channels and interrupts
can have separate security configuration.

This change adds the option to make the GPIOTE20 and
GPIOTE30 secure peripherals and to set a mask which
defines which channels and interrupts should be
configured as secure.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: Ia4422506b0c9af70da252262bcdc205ac8829511