RSE: DCSU prevent locked OTP write
Update DCSU driver to prevent writing to OTP fields locked by a zero
count.

If an OTP field is locked by a zero count it should not be written to
however if the DCSU received a command to write to a zero counted region
it would allow the write if the field was not already written. This
would cause the next boot to fail so instead return an error status and
prevent the write.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: Ib58ad38f406079950c3aa3130e11ef597f921daa