platform: Add shared data getter functions
Add tfm_platform_get_shared_measurement functions and replace usages of
the SHARED_BOOT_MEASUREMENT macros with these functions. We add a
default implementation of these functions which just returns the
existing macros to avoid any changes in behaviour.

These getter functions will be used to allow platforms to customise the
location of the shared boot data and allow them to determine it
dynamically.

Note that the MCUBOOT_SHARED_DATA macros are not removed as part of this
commit and instead we change the definition to use the new getter
macros.

Change-Id: I41c004053ad3754b0cf585bbe1065e4c9d928a4d
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>