RSE: Add const to OTP pointer functions argument
These functions can be used to get a pointer to the SoC area by
passing a pointer to soc_area_info to the function. This area info
variable is of const type and therefore this leads to warnings.

Change the functions to take a const argument to avoid this issue; it
does not make a difference to existing use-cases as we are not modifying
the area_info structure in either function.

Change-Id: Ifef0463784aa6943cec4b600a366658238e9dfe2
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>