DCSU: Add generic error RX response
Add a new error code for RX messages into the DCSU which specifies a
generic error. This will be used only if an error occurs and none of
the other error criteria are satisfied.

This is used to fix an existing issue, whereby the RSE would respond
with a DCSU_ERROR_RX_MSG_GENERIC_ERROR value in the case that none of
the error codes are satisfied. This is an enum of the type designed for
internal RSE use and not to be sent back over the DCSU, therefore it is
larger than the permitted response code.

Change-Id: Ic30ec95a716b9e7b87cdf27d6486928868c60dfa
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>