DCSU: Add test cases for SoC Area
Add additional test cases to the existing DCSU test suite which verify
the writing to and reading from the SoC Area. These tests verify that
the generate SoC ID command works correctly and that we can write and
read valid values to the SoC family ID and IEEE ECID. There are also
tests cases for the SoC configuration area handling.

Change-Id: I161ffe9b9bf14cb698992a453af6f21c63f329fa
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>