Armclang: Fix shared symbol file generation
Previously, the Armclang generated the shared symbol file by using the
--symdefs argument to the linker and then stripping any unneeded symbols
from the file. The --symdefs argument causes Armclang to generate a list
of symbols, but does not include linker script symbols (Image$$ etc.).
A list of these symbols can only be generated from the 'fromelf'
command.

Therefore instead of generating the symdefs file, use 'fromelf' to
generate a symbol dump and parse this into a similar format to the
symdefs file. We add a new script to do this; this script also matches
against the list of symbols (as was previously done by the oneline
python3 command) and strips those which are not required.

Change-Id: I253fa57ff923fc5b4d12786c235b27f96f2a3e6f
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>