rse: check the alignment of the MHU buffer size
validate_buffer_params() takes buf_size argument but does not use it.
It's likely that IS_ALIGNED(buf_addr, 4) was meant to be
IS_ALIGNED(buf_size, 4) as done in MHU v2.

The MHU v3 receive() / send() APIs only handle aligned sizes to 4 bytes,
therefore it makes sense to check both the address and the size are
4-byte aligned.

Change-Id: I9b9d1601e7170c48492b017929a12610cae41c03
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>