DCSU: Tests: Only execute once
Add a check for whether the SOC UID in the OTP has been written when
adding the DCSU tests to the testsuite. This will prevent the tests from
running more than once. This is required for the DCSU tests as various
tests write to the OTP and will fail to run a second time when
attempting to write to the OTP.

Change-Id: I7ee800141e4fde3706924a18cef123c1ef56a7b6
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>