RSE: Add provisioning message state machine
Add a new enum which specifies the component which has received the
provisioning message which is currently staged. This allows for much
cleaner and more robust logic around handling of blobs in runtime,
BL1_2 and chained etc. as we can always be sure that we are processing
the correct blob.

The component that receives the blob is expected to
set the value to its enum and this will then be reset to NO_MESSAGE on
error or after the blob has been parsed and processed (blob code is
expected to call blob_provisioning_finished to indicate that has
completed its task.

Change-Id: I8c38518e671af4c806a03a38686b05281cbd0083
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>