rse: css-aspen: Increase buffers for EFI vars
The UEFI variables are stored in the Protected Storage. The size of
the variables metadata have been increased so the related buffer sizes
have to be increased.

The new values are chosen from test results. Among our test cases,
SystemReady IR ACS test has the highest requirement on the number of
UEFI variables. The values are the minimal ones to pass the BBSR + ACS
test pass for SetVariable and non-volatile storage tests along with 4
additional variables.

Change-Id: I160a9b8da385934f9c2b268fecca7b1bdc28f223
Signed-off-by: Bence Balogh <bence.balogh@arm.com>
Signed-off-by: Michael Zhao <michael.zhao2@arm.com>
Signed-off-by: Ahmed Azeem <ahmed.azeem@arm.com>
Signed-off-by: Debbie Martin <debbie.martin@arm.com>