rse: generate DTCM random word before enabling TRAM encryption
DTCM is set to a random word after enabling TRAM encryption. However,
this random word is currently generated after the TRAM key is exported,
at this point data within the DTCM can no longer be read as this
requires the previous TRAM key. This data covers the global variables
used by the drivers generating random variables. As a result, an ECC
error would be raised in such case.

To fix this, generate the required random word before enabling TRAM
encryption.

Change-Id: I7490ae704d311ce62d250fd3f87e6b3b85da6fc7
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>