RSE: Split provisioning values
Split the provisioning values structure into separate secret and
non-secret structures. The non-secret data contains all of the CM and DM
areas provisioning data, with the secret area containing any keys which
are mandated to be secret.

This non-secret data is then appended to the data section of the blob
directly after any data from the .data section of the blob elf file. We
update the scripting and blob code to support this change.

Change-Id: I42a7953e1022702310ed029ff3f5005396417cc1
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>