RSE: Add routing table header and generate
Add a new header which defines two structures for the RSE routing
tables. One of these is the local routing tables for a single RSE and
the second structure is for the whole system. Note that the single RSE
routing tables can in theory contain tables for more than just the other
RSEs in the system and in future the NUM_NODES will be extended to
support this.

We also add a new routing_tables.py module in a similar vein to the
other provisioning modules. This contains a class definition and logic
to parse the existing layout (TGF files). This script is then added to
the build system and called to generate a pickle file which will be
consumed by other CMake logic to add these routing tables to the RSE
build.

Change-Id: Ie5998a62d48c3cb1f7f8d7dad9a24bf4150b67e5
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>