RSE: Add system routing table to BL1_1
In the case that the routing tables are not in the OTP, they are
compiled into BL1_1 instead. In order to facilitate this, we add a new
script which parses the Routing_tables python class and uses it to
generate a C source file. This can then be added to the BL1_1 source
files.

Change-Id: Ie6653da8ef00b8553383e500bba0e4f87bc3425a
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>