RSE: Re-add routing tables and rse_id to OTP
Routing tables and rse_id were removed from the provisioning bundle
binary generation as part of the refactor (but the fields remained in
the OTP layout structures).

Use the new structure defining the routing tables for a single RSE in
the OTP layout and populate it using the previously generated pickle
file. This requires the generation of individual bundles for each
separate RSE in the system and therefore we add the CMake logic to
facilitate this. We only change the blob naming in the case unique
bundles are required so this does not impact platforms that do not use
this feature.

Change-Id: I9082ae435114308139c48ffd3bab158f27258896
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>